requirements.txt
pandas==0.24.2
python-dotenv
readchar==4.0.5
Unidecode==1.3.8
